Restructure internal representation of active periods
Now relies on the period_status data structure, which indexes each youth
ID to a new Status object that tracks which periods are used, unused, to
be added, and to be removed.

Completely removed the caching system and profile loading system, which
were not very useful in the first place.

Rewrote large portions of the code to rely on period_status instead of
other various ways to aggregate data.