Display number of pending changes for each apron group
Pass stronger cell_is_status function out of match table, which tells
the exact status rather than just achieved or not.

If there are any pending changes under a given apron color, displays
that in the header as well as number achieved (which now no longer
changes, instead always displaying the number of achieved skills in the
database).

Fixes #159