Extract custom filter function logic to Search.js
Takes some configuration functions as parameters, which the date
functions have been made into.

Basically, takes functions for extracting the specified option value
from a cell and for interpreting the search term, such that the results
of those two can be compared.

Also has each operation specify its own comparison function, that can be
run on the output values specified above to determine whether that
filter passes or not.

The goal is to extract as much logic as possible to Search.js, to make
the custom filter as easy as possible to configure and reuse.