Disambiguate youth's class if also the name of an Apron
Specifically, if the youth's class has the form "___Apron", add
" (Class)" to the end to disambiguate it. Mostly relevant for Black
Apron, which is both an apron level and a class.

Also rearranged it so the four basic pieces of info under the youth's
name are in the same order as their corresponding tabs.