Redesign custom filter dropdown
Dropdown is now its own component that is displayed/hidden based on the
button in the Tabulator column header.

ProfileLog (i.e. parent component) handles communication between
Tabulator and filter dropdown.  Dropdown exposes a few functions through
an emit.

Searching now supports filter groups - can search with AND and OR, so in
theory every possible search should be possible using the dropdown (if not
necessarily practical).  Included popup modal to explain a bit more how
filtering works, though it may need rewording to be more user-friendly.

Centering the dropdown is not yet implemented - currently only aligns with
the left side of the button.

Search operations with zero or two inputs not yet implemented.

Inclusive ranges not yet implemented.