Change from dropdown to popup modal
My biggest concerns with the dropdown are:

- If it gets too tall, the page won't let the user scroll down and
actually use it

- Keeping elements of each filter on one line is not straightforward,
and I want them to stay on one line if at all possible

Using a modal should make it more straightforward to set the filters.
The only downside is that it doesn't feel as connected to the table as
the dropdown did.